Add an option to detect all access methods
lspci and setpci have a command line option "-A help" that lists possible access methods but it does not tell the user if the access methods will work. For example intel-conf2 is an access method that is usually listed but usually does not work.

The user can use the -G option to see the results of the detect function of each access method but it stops after the first success. Therefore the user will usually never see the detect result for intel-conf2 because it is usually the last access method.

This patch adds a "-A detect" command line option which calls the detect function of all access methods. Debug messages are enabled so the user can see the results.

Note: specifying an access method (e.g. "-A intel-conf2") still bypasses the detect functions.

Signed-off-by: Joe van Tunen <joevt@shaw.ca>