Show hidden PCI devices
A functioning PCIe device can hide itself by having FFFF:FFFF for vendor and device ID. This patch will make working with these hidden devices possible. A Thunderbolt 3 controller is an example. It can hide itself until code in firmware or a driver unhides the device.

This feature works only with access methods (such as intel-conf1 and darwin) that allow arbitrary access to PCI config space. Access methods such as linux-sysfs and linux-proc can only access PCI devices that they know about.

Unhiding a Thunderbolt 3 controller requires access to extended capabilities (registers beyond the first 256 bytes) of the hidden device. This can be done with the mm command of UEFI Shell or the darwin access method. The intel-conf1 and intel-conf2 access methods only allows access to the first 256 bytes. linux-sysfs and linux-proc can access extended capabilities (with sudo) but not those of hidden devices. Therefore linux needs a new access method that allows arbitrary access to PCIe config space (like the darwin access method) but that is beyond the scope of this patch.

Signed-off-by: Joe van Tunen <joevt@shaw.ca>